Refactor calculator: JS UI, decimal support, cleanup
- Removed server-side calculator logic from CalcuCont controller.
- Updated HomeController to format results to two decimals and show error messages in ViewBag.result.
- Changed Calc and Calcu models to use decimal types and nullable strings for better precision and flexibility.
- Replaced Actualcalc.cshtml form UI with a JavaScript-based calculator interface, including error handling.
- Allowed decimal input in Calculator.cshtml by removing type="number".
- Added "The Calcu" navigation button to Index.cshtml for new calculator UI.

+++ WebApplication1/Models/Calculate.cs
@@ -4,10 +4,10 @@ namespace Calculator.Models
 {
     public class Calc
     {
-        public int result { get; set; }
+        public decimal result { get; set; }
    
-        public int numerator { get; set; }
+        public decimal numerator { get; set; }
 
-        public int denominator { get; set; }
+        public decimal denominator { get; set; }
     }
 }
